Move metadata to its own subpackage.
