+++ lib/objectstore/metadata/metadata_test.go
@@ -1,4 +1,4 @@
-package objectstore
+package metadata
 
 import "testing"
